[CHORE] refactored classes to make structure more logical
