+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/utils/config/WebAppITConfig.java
@@ -1,14 +1,14 @@
 package org.molgenis.integrationtest.utils.config;
 
-import com.google.gson.Gson;
+import org.molgenis.integrationtest.data.settings.SettingsTestConfig;
 import org.molgenis.integrationtest.ui.UiTestConfig;
 import org.molgenis.ui.MolgenisWebAppConfig;
-import org.molgenis.util.GsonHttpMessageConverter;
+import org.molgenis.util.GsonConfig;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 @Configuration
-@Import({ GsonHttpMessageConverter.class, Gson.class, UiTestConfig.class })
+@Import({ GsonConfig.class, UiTestConfig.class, SettingsTestConfig.class })
 public class WebAppITConfig extends MolgenisWebAppConfig
 {
 }
